requirements.txt
# Eye-in-Hand 项目依赖
numpy>=1.20.0
opencv-python>=4.5.0
pyrealsense2>=2.50.0
scipy>=1.7.0
Pillow>=8.0.0
